fixing bugs

+++ run_full_process.sh
@@ -17,7 +17,11 @@ if [ ! -f "$prefix.train-truth.jsonl" ] ||  [ ! -f "$prefix.test-truth.jsonl" ];
     exit 1
 fi
 
-tmpdir=$(mktemp --tmpdir=/tmp -d "preprocess.XXXXXXXXXX")
+#tmpdir=$(mktemp --tmpdir=/tmp -d "process.XXXXXXXXXX")
+tmpdir=/home/moreaue/caca
+rm -rf $tmpdir
+mkdir $tmpdir
+echo "tmpdir=$tmpdir" 1>&2
 cd $tmpdir
 cp -R "$DIR/helper_functions" .
 cp -R "$DIR/preprocessing" .
@@ -35,25 +39,56 @@ cd ../preprocessing
 
 echo "step 1" 1>&2
 python step1-specified-train-test.py "$prefix.train" "$prefix.test"
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 echo "step 2" 1>&2
 python step2_preprocess.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 echo "step 3" 1>&2
 python step3_count.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 echo "step 4" 1>&2
 python step4_make_vocabularies.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 echo "steps 5-7" 1>&2
 python step5_sample_pairs_cal.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 python step6_sample_pairs_val.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 python step7_sample_pairs_dev.py
-
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 echo "7 steps done, copying preprocessing dir" 1>&2
 rm -rf "$prefix.data_preprocessed"
-mv "$tmpdir/data_preprocessed" "$prefix.data_preprocessed"
+cp -R "$tmpdir/data_preprocessed" "$prefix.data_preprocessed"
 
 echo "training model, part 1..." 1>&2
 
 cd $tmpdir/training_adhominem
 python train_adhominem.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 
 echo "copying model 1..." 1>&2
 rm -rf "$prefix.results_adhominem"
@@ -64,6 +99,10 @@ echo "training model, part 2..." 1>&2
 
 cd $tmpdir/training_o2d2
 python train_o2d2.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 
 echo "copying model 2..." 1>&2
 rm -rf "$prefix.results_o2d2"
@@ -73,6 +112,10 @@ cp -R "results_o2d2" "$prefix.results_o2d2"
 echo "inference regular..." 1>&2
 cd $tmpdir/inference
 python run_inference.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 
 echo "getting predicted regular.." 1>&2
 cat predicted.tsv >"$prefix.predicted.regular.tsv"
@@ -83,6 +126,10 @@ echo "inference special ..."
 rm -rf "../results_adhominem" "../results_o2d2"
 cp -R "$DIR/pretrained_models/results_adhominem" "$DIR/pretrained_models/results_o2d2" ..
 python run_inference.py
+if [ $? -ne 0 ]; then
+    echo "ERROR" 1>&2
+    exit 1
+fi
 
 echo "getting predicted special.." 1>&2
 cat predicted.tsv >"$prefix.predicted.special.tsv"
